+++ src/libtai/caltime_tai.c
@@ -17,7 +17,7 @@ void caltime_tai(struct caltime *ct, struct tai *t)
   s = ct->hour * 60 + ct->minute;
   s = (s - ct->offset) * 60 + ct->second;
 
-  t->x = day * 86400ULL + 4611686014920671114ULL + (int64_t) s;
+  t->x = day * ULL(86400) + ULL(4611686014920671114) + (int64_t) s;
 
   leapsecs_add(t,ct->second == 60);
 }
