+++ src/libtai/taia_pack.c
@@ -8,13 +8,13 @@ void taia_pack(char *s, struct taia *t)
   s += 8;
 
   x = t->atto;
-  s[7] = x & 255; x >>= 8;
-  s[6] = x & 255; x >>= 8;
-  s[5] = x & 255; x >>= 8;
-  s[4] = x;
+  s[7] = (int)x & 255; x >>= 8;
+  s[6] = (int)x & 255; x >>= 8;
+  s[5] = (int)x & 255; x >>= 8;
+  s[4] = (int)x;
   x = t->nano;
-  s[3] = x & 255; x >>= 8;
-  s[2] = x & 255; x >>= 8;
-  s[1] = x & 255; x >>= 8;
-  s[0] = x;
+  s[3] = (int)x & 255; x >>= 8;
+  s[2] = (int)x & 255; x >>= 8;
+  s[1] = (int)x & 255; x >>= 8;
+  s[0] = (int)x;
 }
